Extract deprecated class hashing to its own module
