+++ src/starknet/business_logic/fact_state/mod.rs
@@ -1,1 +1,2 @@
 pub mod contract_class_objects;
+pub mod deprecated_class_hash;
